Adding Main Menu and fixes

+++ Assets/NicholasWork/AlienMovement.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class AlienMovement : MonoBehaviour
 {
     // Start is called once before the first execution of Update after the MonoBehaviour is created
-    public int movementSpeed = 5f; 
+    public float movementSpeed = 5f; 
     
     void Start()
     {
